refactor previous homevisit immunization code.
